backend/pyproject.toml
[project]
name = "talent-sourcing-backend"
version = "0.1.0"
description = "Backend for the Talent Sourcing Web Application"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "fastapi>=0.104.0,<0.105.0",
    "uvicorn>=0.24.0",
    "motor>=3.3.2",
    "pymongo>=4.6.1",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.2.1",
    "PyPDF2>=3.0.1",
    "openai>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "httpx>=0.25.1",
    "black>=23.10.1",
    "isort>=5.12.0",
    "mypy>=1.6.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["app/tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v -s --tb=short"
asyncio_mode = "auto"

[tool.black]
line-length = 88
target-version = ["py311"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88 